backend/requirements.txt

Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
psycopg2-binary
whitenoise==6.7.0
